Move rename KPAOption as PodAutoscalerOption.
